Speed up pow(+(...),n).expand() where n>2.
    This new algorithm is faster than the old one because it re-uses
    multinomial coefficients as much as possible.

From GiNaC commit d56a0f74afa5380a1730599c3b1b21f34be2f061
This is the second attempt. It reintroduces #98 and other bugs
which we'll try to fix after this.

+++ src/sage/symbolic/ginac/power.h
@@ -97,7 +97,7 @@ protected:
 	void do_print_python(const print_python & c, unsigned level) const;
 	void do_print_python_repr(const print_python_repr & c, unsigned level) const override;
 
-	ex expand_add(const add & a, int n, unsigned options) const;
+	ex expand_add(const add & a, long int n, unsigned options) const;
 	ex expand_add_2(const add & a, unsigned options) const;
 	ex expand_mul(const mul & m, const numeric & n, unsigned options, bool from_expand = false) const;
